Move fetch patching code to @rsc-parser/core
Created a `fetchPatcher` function in `@rsc-parser/core` that has a callback function called `onRscChunkMessage`. This can then be used in both `@rsc-parser/embedded` and `@rsc-parser/chrome-extension`.

I've also decided to remove `fetchHeaders` from RscChunkMessage for now since it never got used.
