+++ .changeset/plenty-dingos-jam.md
@@ -1,0 +1,10 @@
+---
+"@rsc-parser/chrome-extension": minor
+"@rsc-parser/embedded": minor
+"@rsc-parser/core": minor
+"@rsc-parser/embedded-example": minor
+"@rsc-parser/storybook": minor
+"@rsc-parser/website": minor
+---
+
+Move fetch patching code to @rsc-parser/core
