+++ packages/core/src/example-data/alvar-dev.ts
@@ -5,12 +5,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://www.alvar.dev/blog?_rsc=1sxgj",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526298,
       chunkValue: [
         50, 58, 73, 91, 50, 53, 50, 53, 48, 44, 91, 34, 53, 50, 53, 48, 34, 44,
@@ -34,12 +28,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://www.alvar.dev/blog?_rsc=1sxgj",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526298,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -457,12 +445,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://www.alvar.dev/blog?_rsc=1sxgj",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526298,
       chunkValue: [
         52, 58, 91, 34, 36, 34, 44, 34, 100, 105, 118, 34, 44, 110, 117, 108,
@@ -754,13 +736,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/thoughts-on-photography-tools?_rsc=18e5w",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2Cnull%2Ctrue%5D",
-        "Next-Router-Prefetch": "1",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526548,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -788,13 +763,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/always-add-name-to-type-radio?_rsc=18e5w",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2Cnull%2Ctrue%5D",
-        "Next-Router-Prefetch": "1",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526581,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -822,13 +790,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/tailwindcss-with-next-font?_rsc=18e5w",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2Cnull%2Ctrue%5D",
-        "Next-Router-Prefetch": "1",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526547,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -856,13 +817,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/creating-devtools-for-react-server-components?_rsc=18e5w",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2Cnull%2Ctrue%5D",
-        "Next-Router-Prefetch": "1",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526547,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -891,13 +845,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/skeleton-loading-with-suspense-in-next-js-13?_rsc=18e5w",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(main)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2Cnull%2Ctrue%5D",
-        "Next-Router-Prefetch": "1",
-        "Next-Url": "/blog",
-      },
       fetchStartTime: 1706448526547,
       chunkValue: [
         48, 58, 91, 34, 87, 89, 106, 69, 45, 79, 83, 72, 85, 105, 48, 45, 87,
@@ -926,12 +873,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/creating-devtools-for-react-server-components?_rsc=1yujw",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(backnav)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%5B%22slug%22%2C%22creating-devtools-for-react-server-components%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D",
-        "Next-Url": "/blog/creating-devtools-for-react-server-components",
-      },
       fetchStartTime: 1706448528308,
       chunkValue: [
         51, 58, 73, 91, 53, 54, 49, 51, 44, 91, 93, 44, 34, 34, 93, 10, 53, 58,
@@ -1255,12 +1196,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/creating-devtools-for-react-server-components?_rsc=1yujw",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(backnav)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%5B%22slug%22%2C%22creating-devtools-for-react-server-components%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D",
-        "Next-Url": "/blog/creating-devtools-for-react-server-components",
-      },
       fetchStartTime: 1706448528308,
       chunkValue: [
         56, 58, 73, 91, 56, 49, 57, 55, 50, 44, 91, 34, 53, 50, 53, 48, 34, 44,
@@ -3728,12 +3663,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/creating-devtools-for-react-server-components?_rsc=1yujw",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(backnav)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%5B%22slug%22%2C%22creating-devtools-for-react-server-components%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D",
-        "Next-Url": "/blog/creating-devtools-for-react-server-components",
-      },
       fetchStartTime: 1706448528308,
       chunkValue: [
         49, 50, 58, 34, 36, 76, 49, 55, 34, 10, 49, 51, 58, 34, 36, 76, 49, 56,
@@ -3750,12 +3679,6 @@ export const alvarDevExampleData: RscChunkMessage[] = [
     data: {
       fetchUrl:
         "https://www.alvar.dev/blog/creating-devtools-for-react-server-components?_rsc=1yujw",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(backnav)%22%2C%7B%22children%22%3A%5B%22blog%22%2C%7B%22children%22%3A%5B%5B%22slug%22%2C%22creating-devtools-for-react-server-components%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D",
-        "Next-Url": "/blog/creating-devtools-for-react-server-components",
-      },
       fetchStartTime: 1706448528308,
       chunkValue: [
         49, 55, 58, 91, 34, 36, 34, 44, 34, 100, 105, 118, 34, 44, 110, 117,
