+++ packages/core/src/example-data/gh-fredkiss-dev.ts
@@ -5,12 +5,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=kvatu",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203109,
       chunkValue: [
         55, 58, 73, 91, 50, 48, 55, 48, 44, 91, 93, 44, 34, 34, 93, 10, 97, 58,
@@ -39,12 +33,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=kvatu",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203109,
       chunkValue: [
         48, 58, 91, 34, 103, 90, 86, 119, 81, 120, 87, 56, 117, 51, 50, 115, 71,
@@ -462,12 +450,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=kvatu",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203109,
       chunkValue: [
         50, 58, 91, 34, 36, 34, 44, 34, 110, 97, 118, 34, 44, 110, 117, 108,
@@ -655,12 +637,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=1326h",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203252,
       chunkValue: [
         51, 58, 73, 91, 50, 48, 55, 48, 44, 91, 93, 44, 34, 34, 93, 10, 54, 58,
@@ -1056,12 +1032,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=1326h",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203252,
       chunkValue: [
         55, 58, 91, 91, 34, 36, 34, 44, 34, 109, 101, 116, 97, 34, 44, 34, 48,
@@ -1287,12 +1257,6 @@ export const ghFredkissDevExampleData: RscChunkMessage[] = [
     type: "RSC_CHUNK",
     data: {
       fetchUrl: "https://gh.fredkiss.dev/Fredkiss3/gh-next/actions?_rsc=1ho2t",
-      fetchHeaders: {
-        RSC: "1",
-        "Next-Router-State-Tree":
-          "%5B%22%22%2C%7B%22children%22%3A%5B%22(app)%22%2C%7B%22header_subnav%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22page_title%22%3A%5B%22children%22%2C%7B%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22%5B...pages%5D%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%5D%7D%5D%7D%5D%7D%5D%2C%22children%22%3A%5B%5B%22user%22%2C%22Fredkiss3%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%5B%22repository%22%2C%22gh-next%22%2C%22d%22%5D%2C%7B%22children%22%3A%5B%22actions%22%2C%7B%22children%22%3A%5B%22__PAGE__%22%2C%7B%7D%5D%7D%2Cnull%2C%22refetch%22%5D%7D%5D%7D%5D%7D%5D%7D%5D",
-        "Next-Url": "/children/Fredkiss3/gh-next/[...pages]",
-      },
       fetchStartTime: 1706451203373,
       chunkValue: [
         57, 58, 73, 91, 57, 57, 49, 52, 44, 91, 34, 57, 57, 49, 52, 34, 44, 34,
