+++ packages/core/src/types.ts
@@ -3,7 +3,6 @@ export type RscChunkMessage = {
   tabId: number;
   data: {
     fetchUrl: string;
-    fetchHeaders: Record<string, string | undefined>;
     fetchStartTime: number;
     chunkValue: number[];
     chunkStartTime: number;
